fix: rename column table
