+++ src/components/UserAccount/Collections/CollectionTable.jsx
@@ -6,6 +6,7 @@ const CollectionTable = () => {
       <div className="flex justify-end items-center gap-3">
         <CollectionButtons />
       </div>
+
     </section>
   );
 };
